# back-end/docs/api/swagger.yaml
definitions:
  apperrors.AppError:
    properties:
      http_status:
        type: integer
      message:
        type: string
    type: object
...
  model.ChangeEmailRequest:
    properties:
      code:
        type: string
      email:
        type: string
    type: object
...
  model.User:
    properties:
      created_at:
        example: "2025-10-01T03:12:20Z"
        type: string
      email:
        example: email@example.com
        type: string
      id:
        example: 1234567890
        type: integer
      public_key:
        type: string
      role_id:
        example: 1
        type: integer
      updated_at:
        example: "2025-10-01T03:12:20Z"
        type: string
    type: object
  model.UserAction:
    enum:
    - add_api
    - delete_api
    - wallet_deposit
    - wallet_import
    - wallets_batch_create
    type: string
    x-enum-varnames:
    - ActionAddAPI
    - ActionDeleteAPI
    - ActionWalletDeposit
    - ActionWalletImport
    - ActionWalletsBatchAdd
  model.UserHistory:
    properties:
      action:
        $ref: '#/definitions/model.UserAction'
      created_at:
        type: string
      id:
        type: integer
      user_id:
        type: integer
      value:
        type: string
    type: object
...
host: https://api-stage.tsunammi.io/
info:
  contact: {}
  description: This is the swagger specification for Tsunammi API.
  title: Tsunammi API
  version: 0.0.7
paths:
...
  /user:
    get:
      description: Returns the profile data for the user authenticated via a JWT access
        token.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            $ref: '#/definitions/model.User'
        "401":
          description: Authentication error, claims not found
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: Get current user info
      tags:
      - user
  /user/email:
    patch:
      consumes:
      - application/json
      description: Updates the email address of the authenticated user. Requires a
        valid confirmation code sent to the new email.
      operationId: update-user-email
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: New email and verification code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ChangeEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Invalid request data or email format
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "401":
          description: Unauthorized or invalid verification code
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: Update user email
      tags:
      - user
  /user/history:
    get:
      description: Returns a list of actions (history) for the authenticated user.
      operationId: get-user-history
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User action history
          schema:
            items:
              $ref: '#/definitions/model.UserHistory'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: Get user action history
      tags:
      - user
...
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer' followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"